vacci_track_backend/requirements.txt
djangorestframework
Django
cx_oracle
psycopg2
django-cors-headers
djangorestframework-simplejwt
sqlalchemy
# uwsgi